refactor: Separate helper functions for data, mocking fetch and file read responses
